Added PythonTestScene. Python for Unity test status: Failed
Scene was created with intent to test 'Python for Unity' Package.
- This package only uses Python 2 and is not compatible with Python 3.
- Python 2 is now 'End of Life' so it is pretty much obsolete.

I therefore decided to implement a simple keyword search in C#.
Basically:
if (this keyword found!)
       output this message.
else if (this keyword found!)
       output this message.
This ensures that no two messages will be output.
This is not efficient obviously but for prototype purposes, and with the AI Implementation in python, this is more than enough.

TO DO: Try and implement a scene check to provide key info
Create GetScene() function
Use GetScene() function to create more personalised AI outputs.
E.G. if GetScene() returns Scene.TeeScene then Sloth says "Hey, Nice Tee! Any questions?"

Also, the first if statement checks for profanity because Sloth don't deserve that...